ponderada_4/image_service/requirements.txt
annotated-types==0.7.0
anyio==4.3.0
attrs==23.2.0
async-timeout==4.0.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
coloredlogs==15.0.1
cryptography==41.0.3
dnspython==2.6.1
email_validator==2.1.1
exceptiongroup==1.2.1
fastapi-cli==0.0.4
fastapi-slim==0.111.0
flatbuffers==24.3.25
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
humanfriendly==10.0
idna==3.7
imageio==2.34.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
lazy_loader==0.4
llvmlite==0.42.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mpmath==1.3.0
networkx==3.3
numba==0.59.1
numpy==1.26.4
onnxruntime==1.18.0
opencv-python-headless==4.9.0.80
orjson==3.10.3
packaging==24.0
pillow==10.3.0
platformdirs==4.2.2
pooch==1.8.1
protobuf==5.27.0
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.18.0
PyJWT==2.8.0
PyMatting==1.1.12
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
redis==5.0.4
referencing==0.35.1
rembg==2.0.57
requests==2.32.2
rich==13.7.1
rpds-py==0.18.1
scikit-image==0.23.2
scipy==1.13.1
shellingham==1.5.4
sniffio==1.3.1
starlette==0.37.2
sympy==1.12
tifffile==2024.5.22
tqdm==4.66.4
typer==0.12.3
typing_extensions==4.12.0
ujson==5.10.0
urllib3==2.2.1
uvicorn==0.29.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
